+++ src/SdCard.h
@@ -3,16 +3,16 @@
 
 #include <SD.h>
 
-const float FILE_CLOSE_TIME = 1;
-const float CARD_CLOSE_TIME = 10;
-const float CHECK_INTERVAL  = 0;
+const float FILE_CLOSE_TIME = 1.0;
+const float CARD_CLOSE_TIME = 10.0;
+const float CHECK_INTERVAL  = 0.0;
 
-const int PIN_SD_CS        = 49; // Chip Select
-const int PIN_SD_CD        = 53; // Card Detect
-const int PIN_SD_LED_RED   = 44; // RGB Red pin
-const int PIN_SD_LED_GREEN = 45; // RGB Green pin
-const int PIN_SD_LED_BLUE  = 46; // RGB Blue pin
-const int SD_INDICATOR_POWER  = 10;
+const uint8_t PIN_SD_CS          = 49; // Chip Select
+const uint8_t PIN_SD_CD          = 53; // Card Detect
+const uint8_t PIN_SD_LED_RED     = 44; // RGB Red pin
+const uint8_t PIN_SD_LED_GREEN   = 45; // RGB Green pin
+const uint8_t PIN_SD_LED_BLUE    = 46; // RGB Blue pin
+const uint8_t SD_INDICATOR_POWER = 10;
 
 class SdCard
 {
@@ -43,7 +43,7 @@ private:
 	File logFile;
 	boolean logFileAvailable, fileOpened, cardOpened, cardPresent;
 	char sdFileName[13];
-	unsigned long fileLastOpened, cardLastChecked;
+	uint32_t fileLastOpened, cardLastChecked;
 	boolean enabled, tempDisabled;
 
 	boolean createLogFile ();
