+++ src/Utils.h
@@ -11,9 +11,9 @@
 // ************************************* //
 
 // Caps
-const int CAPS_NONE  = 0; // All letters in lower case
-const int CAPS_FIRST = 1; // First letter in upper case
-const int CAPS_ALL   = 2; // All letters in upper case
+const uint8_t CAPS_NONE  = 0; // All letters in lower case
+const uint8_t CAPS_FIRST = 1; // First letter in upper case
+const uint8_t CAPS_ALL   = 2; // All letters in upper case
 
 const uint8_t COMPLEMENT_TYPE_NONE = 0; // No complement
 const uint8_t COMPLEMENT_TYPE_LMO  = 1; // Light mod
@@ -24,8 +24,8 @@ const uint8_t COMPLEMENT_TYPE_SCP  = 2; // Sound command parameter
 // ************************************************************* //
 
 // Min and Max
-const int SEEKBAR_MIN = 0;   // Minimum app seek bars value
-const int SEEKBAR_MAX = 100; // Maximum app seek bars value
+const uint8_t SEEKBAR_MIN = 0;   // Minimum app seek bars value
+const uint8_t SEEKBAR_MAX = 100; // Maximum app seek bars value
 
 class Utils
 {
@@ -39,8 +39,8 @@ public:
 	const uint8_t infoTypeComplementType (uint8_t infoType);
 	char * clock (char * buf); // a 20-char buf is necessary
 	void softwareReset ();     // Just in case
-	char * reduceCharArray (char ** array, int length);
-	long map (float input, float inMin, float inMax, float outMin, float outMax);
+	char * reduceCharArray (char ** array, uint16_t length);
+	uint32_t map (float input, float inMin, float inMax, float outMin, float outMax);
 };
 
 extern Utils utils;
